renames component
