+++ src/frontend/src/lib/components/nfts/NftSettingsMenu.svelte
@@ -7,7 +7,7 @@
 	import IconWarning from '$lib/components/icons/IconWarning.svelte';
 	import IconEyeOff from '$lib/components/icons/lucide/IconEyeOff.svelte';
 	import IconManage from '$lib/components/icons/lucide/IconManage.svelte';
-	import NftCollectionShowSpamToggle from '$lib/components/nfts/NftCollectionShowSpamToggle.svelte';
+	import NftsShowSpamToggle from '$lib/components/nfts/NftsShowSpamToggle.svelte';
 	import ButtonIcon from '$lib/components/ui/ButtonIcon.svelte';
 	import LogoButton from '$lib/components/ui/LogoButton.svelte';
 	import ResponsivePopover from '$lib/components/ui/ResponsivePopover.svelte';
@@ -94,7 +94,7 @@
 						<span class="text-sm font-normal">{$i18n.nfts.text.show_spam}</span>
 					{/snippet}
 					{#snippet action()}
-						<NftCollectionShowSpamToggle />
+						<NftsShowSpamToggle />
 					{/snippet}
 				</LogoButton>
 			</ListItem>
