+++ src/frontend/src/lib/components/nfts/NftsShowSpamToggle.svelte
